# Hash function of Polygon class changed