--- a/sympy/geometry/polygon.py
+++ b/sympy/geometry/polygon.py
@@ -1038,41 +1038,24 @@ class Polygon(GeometrySet):
 
     def _hashable_content(self):
 
+        D = {}
         def ref_list(point_list):
             kee = {}
             for i, p in enumerate(ordered(set(point_list))):
                 kee[p] = i
+                D[i] = p
             return [kee[p] for p in point_list]
 
-        S1 = ref_list( self.args)
-        r_nor = rotate_left( S1, least_rotation( S1))
-        S2 = ref_list( list( reversed(self.args) ))
-        r_rev = rotate_left( S2, least_rotation( S2))
+        S1 = ref_list(self.args)
+        r_nor = rotate_left(S1, least_rotation(S1))
+        S2 = ref_list(list(reversed(self.args)))
+        r_rev = rotate_left(S2, least_rotation(S2))
         if r_nor < r_rev:
-            res = r_nor
-        else :
-            res = r_rev
-        return tuple(res)
-
-    def _hashable_content(self):
-        def ref_list(point_list):
-            '''
-            Converts list of Point objects to their corresponding
-            order in sorting.
-            '''
-            kee = {}
-            for i, p in enumerate(ordered(set(point_list))):
-                kee[p] = i
-            return [kee[p] for p in point_list]
-
-        # find minimal rotation of both normal and reverse list
-        # of points and use smaller of the two to compute hash.
-        S1 = ref_list( self.args)
-        r_nor = rotate_left( S1, least_rotation( S1))
-        S2 = ref_list( list( reversed(self.args) ))
-        r_rev = rotate_left( S2, least_rotation( S2))
-        r = r_nor if r_nor < r_rev else r_rev
-        return tuple(r)
+            r = r_nor
+        else:
+            r = r_rev
+        canonical_args = [ D[order] for order in r ]
+        return tuple(canonical_args)
 
     def __contains__(self, o):
         """