Before this commit, hash of two Polygon having same vertices but initialised with
different order were different. This was conceptually wrong.
Now hash of two Polygons having same vectices will be equal irrespective of the
order at time of initialisation.
Hash of Polygons which compare equal 
<!-- Your title above should be a short description of what
was changed. Do not include the issue number in the title. -->

#### References to other Issues or PRs
<!-- If this pull request fixes an issue, write "Fixes #NNNN" in that exact
format, e.g. "Fixes #1234". See
https://github.com/blog/1506-closing-issues-via-pull-requests .-->
Fixes 7998

#### Brief description of what is fixed or changed
Two Polygons are equal if they have set of vertices (AS defined in geometry module). 
Before this PR, hash of two equal Polygons  having different order of vertices was different.
This has been fixed now.
Ex-
`
>>> a = Polygon( Point(0, 0), Point(1, 0), Point(1,1)
>>> b = Polygon( Point(0, 0), Point(1, 1), Point(1, 0)
>>> hash(a) == hash(b)
>>> True
`
#### Other comments
See [#15747](https://github.com/sympy/sympy/pull/15747l)

#### Release Notes

<!-- Write the release notes for this release below. See
https://github.com/sympy/sympy/wiki/Writing-Release-Notes for more information
on how to write release notes. The bot will check your release notes
automatically to see if they are formatted correctly. -->

<!-- BEGIN RELEASE NOTES -->
* geometry
   * hash function of geometry modified
<!-- END RELEASE NOTES -->