@smichr 2019-01-09
How about at instantiation use the rotation of points that would give the lowest ordered point first and in case of a tie, recursively use the point that has the smallest next point, so (1,1),(2,0),(0,2) would be rotated to put (0,2) first: (0,2)(1,1)(2,0) and for (1,1),(0,2),(2,0),(0,2) it would be (0,2),(1,1),(0,2),(2,0) since (1,1) is less than (2,0). This is an O(n) test at instantiation:

```
small=points
for i in range(len(points)-1):
    r = rotate_right(points, 1)
    if next(ordered(small, r)) == r:
        small = r
points = small
```