@asmeurer 2019-01-09
Shouldn't two polygons either have the same vertices in cyclic order or reverse cyclic order to be equal? For instance, `(0, 0), (0, 1), (1, 1), (1, 0)` and `(0, 0), (1, 1), (1, 0), (0, 1)` are not equal (one is a square and one is a polygon with intersecting edges).