@FrackeR011 2019-01-10
I had a wrong idea of equality in my mind. Sortingwas a bad idea. It now gives same hash even for different Polygons as mentioned above.
But i think these Polygons(Line segment) should be compared equal.
```
>>> p1 = Polygon( Point(1, 1), Point(2, 0), Point(0, 2))
>>> p2 = Polygon( Point(1, 1), Point(0, 2), Point(2, 0))
>>> p1
Segment2D(Point2D(1, 1), Point2D(0, 2))
>>> p2
Segment2D(Point2D(1, 1), Point2D(2, 0))
>>> p1==p2
False
```
I think Both p1 and p2 should be `Segment2D(Point2D(0, 2), Point2D(2, 0))`.