@FrackeR011 2019-01-10
I think @smichr  method will work if use both rotate right and rotate left and then choose the smaller among the list