@smichr 2019-01-10
> Can you clarify what `ordered(small, r)` returns

sorry, it should be `next(ordered((small, r)))` is should return the smaller argument in terms of canonical sorting:
```
>>> next(ordered(([(1, 1), (1, 0), (0, 0)],[(0, 0), (1, 0), (1, 1)])))
[(0,0),(1,0),(1,1)]
```