@smichr 2019-01-10
> i believe will have O(n^2) complexity.

you are right.