@FrackeR011 2019-01-10
How about this?
'''
        def compare_list( a ,b ):
            for i in range( len(a) ) :
                if( a[i].args < b[i].args ):
                    return True
                elif( a[i].args > b[i].args):
                    return False
            return False

        small = args
        r = args

        for i in range( len( args) ):
            r = rotate_list(r, 1)
            if compare_list( r , small):
                small = r
        
        r = list( reversed( args))
        for i in range( len(args) ):
            r = rotate_list(r, 1)
            if compare_list( r , small):
                small = r

        points = small
'''
We will rotate point list and its reverse and find the smallest rotation between both?