@smichr 2019-01-10
> if( a[i].args < b[i].args )

This won't work when there are symbolic elements in the Polygon since this will not evaluate to True or False in general.

Can anything in the combinatorics help? If the points are ordered and assigned ints, is there a combinatorics function that will tell you which is the smallest canonical rotation? Hey google, "lexicographical rotations" : [answer is here](https://en.wikipedia.org/wiki/Lexicographically_minimal_string_rotation).