@FrackeR011 2019-01-10
Thanks for pointing out, i will try implementing Booth algorithm( I do not know much about it except what Wikipedia says).
The main difficulty is the list consisting of objects of Point type rather than number of symbols. I will try converting them into strings, then either Booth Algorithm or above algorithm will work.
Anyway, Booth Algorithm is faster and should be used.