@asmeurer 2019-01-10
Equal terms in the ordering should be rare. I don't think you need to be particularly efficient about it.