@smichr 2019-01-11
It would be good to put the `least_rotation` into utilities.iterables.