@jksuom 2019-01-11
It seems to me that `_hashable_content` could simply return `sorted(self.args)`. It does not matter that the cyclic order of the argument points may not be preserved as `hash` is only the first quick step before calling `__eq__`. It is expected that  `__eq__` would do more refined comparison.