@smichr 2019-01-11
> as `hash` is only the first quick step

I don't understand the significance of hash well. I know that I have encountered unexpected results when working with expressions where a previously cached value is presented instead of the one that I was expecting and I thought this was related to the hash value. I could be wrong and it is perhaps better to let others continue with this discussion. Does hash only tell you things that *might* be the same or different? From the issue page,
```
from the Python docs:
__hash__() should return an integer. The only required property is that objects which compare 
equal have the same hash value;
```
From that it doesn't mean that if they have the same hash value that they *are* equal, so perhaps your solution is best. How would having a more discriminating hash -- being different for different permutations of points -- be better or worse than what you propose?