@jksuom 2019-01-11
>  Does hash only tell you things that might be the same or different?

I think it works like this: If it has to be decided whether two objects are same or not (for example, when looking for an object in a list or set), Python first compares their hash values. If they are different, the objects must be different and no more processing is needed, otherwise, `__eq__` has to be called.