@asmeurer 2019-01-11
Hash collisions are fine and self intersecting polygons should be rare anyway. My only worry is there might be code that assumes the _hashable_content itself should be unique.