@FrackeR011 2019-01-12
The nature of hash should be same for every class of sympy. I am in support of same hash for different permutations of polygon.
Do we need to keep current behavior of == as it is or change it as discussed?