@smichr 2019-01-12
> I am in support of same hash for different permutations of polygon.

So are you supporting @jksuom's idea?



> keep current behavior of == as it is

I think I would be in favor of `_eq_` respecting the orientation since area does so; if the objects don't have the same area they shouldn't be equal. Actually, in keeping with SymPy standards, the equality should not produce True unless the two objects are structurally identical. That's why I am in favor of making the objects canonical in terms of rotation at the outset; `equals` *could perhaps* be made to ignore orientation. 

But let's not get bogged down by trying to do it all in one PR. Fix the hash according to the consensus and start afresh (if you want) on the next aspect.