@FrackeR011 2019-01-15
In my system ( Ubuntu 16.04 , python 2.7 )

Master branch- 
``
In [5]: %time p7 = Polygon(Point(x, y), Point(q, u), Point(v, w))
CPU times: user 2.75 s, sys: 11.6 ms, total: 2.76 s
Wall time: 2.76 s
``
This branch ( fix_hash_7998) -
``I
n [6]: %time p7 = Polygon(Point(x, y), Point(q, u), Point(v, w))
CPU times: user 2.74 s, sys: 7.87 ms, total: 2.74 s
Wall time: 2.74 s
``
@jksuom Can you test master branch in your system ?
When numbers are used instead of symbols object creation is much faster which is expected.
``
In [4]: %time p7 = Polygon(Point(0, 0), Point(1, 0), Point(1, 1))
CPU times: user 4.33 ms, sys: 61 µs, total: 4.39 ms
Wall time: 3.75 ms
``
Maybe we need to work on ` __new__` ?